fix: replace broken test and undo renaming
