+++ src/Dfe.PlanTech.Application/Constants/DatabaseConstants.cs
@@ -4,5 +4,5 @@ public static class DatabaseConstants
 {
     public const string CalculateMaturitySprocParam = "@submissionId";
     public const string CalculateMaturitySproc = $"[dbo].[calculateMaturityForSubmission] {CalculateMaturitySprocParam}";
-    public const string GetSectionStatusesForCategory = $"[dbo].[GETSectionStatusesForCategoryKGTest]";
+    public const string GetSectionStatuses = $"[dbo].[GetSectionStatusesForCategory]";
 }
