+++ src/Dfe.PlanTech.Application/Persistence/Interfaces/IPlanTechDbContext.cs
@@ -25,7 +25,7 @@ public interface IPlanTechDbContext
 
     Task<int> CallStoredProcedureWithReturnInt(string sprocName, IEnumerable<SqlParameter> parms, CancellationToken cancellationToken = default);
     
-    IQueryable<SectionStatusDto> GetSectionStatusesForCategory(string categoryId, int establishmentId);
+    IQueryable<SectionStatusDto> GetSectionStatuses(string categoryId, int establishmentId);
 
     Task<User?> GetUserBy(Expression<Func<User, bool>> predicate);
